+++ tests/tests_api_v1_1/test_api_v1_1_serializer_required_fields.py
@@ -17,10 +17,10 @@ class SerializerRequiredFieldsTestCase(APITestCase):
     This model has the following fields with the following constraints:
         - field 'name' is required and cannot be NULL
         - field 'description' is not required
+        - field 'cost' is required
+        - field 'nb_articles' is required
         - field 'status' is not required BUT if found, cannot be NULL
-        - field 'nb_articles' is required and cannot be NULL
-        - field 'cost' is not required BUT if found, cannot be NULL
-        - field 'sold' is not required BUT if found, cannot be NULL
+        - field 'sold' is required
 
     The errors that may occur are the following:
         - ['This field is required.'] => 'required' validation
@@ -54,19 +54,15 @@ class SerializerRequiredFieldsTestCase(APITestCase):
                 'name': '',
                 'description': '',
                 'status': '',
-                'cost': 0.0,
-                'nb_articles': 0,
+                'cost': 10.0,
+                'nb_articles': 10,
+                'sold': True,
             },
             HTTP_AUTHORIZATION=f'Token {self.token}',
         )
         self.assertEqual(resp.status_code, status.HTTP_400_BAD_REQUEST)
         self.assertDictEqual(
-            resp.json(),
-            {
-                'name': ['This field may not be blank.'],
-                'cost': ['This field may not be blank.'],
-                'nb_articles': ['This field may not be blank.'],
-            },
+            resp.json(), {'name': ['This field may not be blank.']}
         )
 
         #:  The header 'application/json' allows to pass empty strings
@@ -78,9 +74,9 @@ class SerializerRequiredFieldsTestCase(APITestCase):
                     'name': '',
                     'description': '',
                     'status': '',
-                    'cost': 0.0,
+                    'cost': None,
                     'sold': None,
-                    'nb_articles': 0,
+                    'nb_articles': None,
                 }
             ),
             content_type="application/json",
@@ -93,9 +89,9 @@ class SerializerRequiredFieldsTestCase(APITestCase):
             {
                 'name': ['This field may not be blank.'],
                 'status': ['This field may not be blank.'],
-                'cost': ['This field may not be blank.'],
+                'cost': ['This field may not be null.'],
                 'sold': ['This field may not be null.'],
-                'nb_articles': ['This field may not be blank.'],
+                'nb_articles': ['This field may not be null.'],
             },
         )
 
@@ -106,13 +102,12 @@ class SerializerRequiredFieldsTestCase(APITestCase):
         self.assertEqual(resp.status_code, status.HTTP_400_BAD_REQUEST)
         self.assertNotIn('description', resp.data)
         self.assertNotIn('status', resp.data)
-        self.assertNotIn('cost', resp.data)
-        self.assertNotIn('sold', resp.data)
         self.assertDictEqual(
-            resp.data,
+            resp.json(),
             {
                 'name': ['This field is required.'],
                 'nb_articles': ['This field is required.'],
+                'cost': ['This field is required.'],
             },
         )
         resp = self.client.post(
@@ -124,13 +119,13 @@ class SerializerRequiredFieldsTestCase(APITestCase):
         self.assertEqual(resp.status_code, status.HTTP_400_BAD_REQUEST)
         self.assertNotIn('description', resp.data)
         self.assertNotIn('status', resp.data)
-        self.assertNotIn('cost', resp.data)
-        self.assertNotIn('sold', resp.data)
         self.assertDictEqual(
             resp.data,
             {
                 'name': ['This field is required.'],
                 'nb_articles': ['This field is required.'],
+                'cost': ['This field is required.'],
+                'sold': ['This field is required.'],
             },
         )
 
@@ -138,13 +133,20 @@ class SerializerRequiredFieldsTestCase(APITestCase):
         self.assertEqual(ItemPack.objects.count(), 0)
         resp = self.client.post(
             self.url,
-            data={'name': 'Test', 'nb_articles': 20},
+            data={
+                'name': 'Test',
+                'nb_articles': 20,
+                'cost': 10.0,
+                'sold': True,
+            },
             HTTP_AUTHORIZATION=f'Token {self.token}',
         )
         self.assertEqual(resp.status_code, status.HTTP_201_CREATED)
         self.assertEqual(ItemPack.objects.count(), 1)
         item_pack = ItemPack.objects.first()
+        self.assertEqual(item_pack.name, 'Test')
+        self.assertEqual(item_pack.nb_articles, 20)
+        self.assertEqual(item_pack.cost, Decimal(10))
+        self.assertEqual(item_pack.sold, True)
         self.assertEqual(item_pack.status, 'PENDING')  # default value
-        self.assertEqual(item_pack.sold, False)  # default value
-        self.assertEqual(item_pack.cost, Decimal(10))  # default value
         self.assertEqual(item_pack.description, '')  # default value