+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 
 ### Changed
 
-- changed serializer fields validations
+- In CharField and TextField serializers, the field is required only if `blank` is `False` AND default value is not an empty string. Otherwise it is not required.
 
 ### Removed